Clean up inputs exposure, stateVersion
As inputs multiply we need more explicitly named inputs to pass to
specialArgs along from the flake outputs signature declaration. This
sucks and is unnecessary, since inputs is already implicitly passed, so
fix this.

Nix state version was a parameter passed as an input before, due to
desire to sync home manager and nixOS state versions, but this is not
required.